% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot results from a Bayesian EWI model}
\usage{
plot_estimates(fitted, alpha_ci = 0.05, link_space = FALSE)
}
\arguments{
\item{fitted}{Object returned from fit_ewi}

\item{alpha_ci}{alpha level for credible intervals, defaults to 0.05 = 95% CIs}

\item{link_space}{Whether to plot the parameters in link or normal space. Defaults to FALSE (normal space)}
}
\description{
Plot results from a Bayesian EWI model
}
\examples{
\dontrun{
library(bayesewi)
model_1 = fit_ewi(data_example, ewi_model="ar", time="continuous")
print(plot_estimates(model_1, alpha_ci = 0.1)) # make plot in normal space
print(plot_estimates(model_1, alpha_ci = 0.1, link_space=TRUE)) # make plot in logit space
}
}
