% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ewi.R
\name{fit_ewi}
\alias{fit_ewi}
\title{Fit a Bayesian EWI model}
\usage{
fit_ewi(data_frame, ewi_model = "ar", time = "discrete", iter = 2000,
  chains = 4, control = list(adapt_delta = 0.9, max_treedepth = 20),
  ar_scale = 0.1)
}
\arguments{
\item{data_frame}{Data frame containing the time covariarte ('x') and response ('y') with no NAs}

\item{time}{Whether to fit model with discrete time ('discrete') or continuous time ('continuous'). Defaults to 'discrete'.}

\item{iter}{Number of iterations in Stan sampling.}

\item{chains}{Number of chains in Stan sampling.}

\item{control}{A list of options to pass to Stan sampling.}

\item{model}{Which model to fit. Either the dynamic AR(1) model ('ar') or stochastic volatility model with time varying volatility ('sv'). Defaults to 'ar'.}
}
\description{
Fit a Bayesian EWI model
}
\examples{
\dontrun{
library(bayesewi)
model_1 = fit_ewi(data_example, ewi_model="ar", time="continuous")
model_2 = fit_ewi(data_example, ewi_model="ar", time="discrete", iter = 1000, chains=2)

print(plot_estimates(model_2))
model_3 = fit_ewi(data_example, ewi_model="sv", time="discrete", iter = 2000, chains=3)
}
}
